pipx

+++ build-scripts/37-as_root-symlink_fonts.sh
@@ -3,6 +3,8 @@
 
 # bash 38-as_root-symlink_fonts.sh
 
+if [[ $(id -u) > 0 ]]; then echo "Run this as root!"; exit; fi
+
 #=> for fonts
 cfA=/usr/share/fonts/copiedForArch
 [[ $cfA ]] && rm -r $cfA; ln -s $IThandy/fonts-forArch $cfA
